+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.3] - 2024-11-25
 
 ### Added
 
 - PR template
+- Validation to options form
+
+### Changed
+
+- FIX typo in marketing cookie text
+- Analytics options are trimmed before adding to HTML output
 
 ## [1.5.2] - 2024-07-08